chore: move getBlockTypeCounts out of utils.js
