+++ core/block.js
@@ -17,9 +17,9 @@ goog.module('Blockly.Block');
 
 /* eslint-disable-next-line no-unused-vars */
 const Abstract = goog.requireType('Blockly.Events.Abstract');
-const arrayUtils = goog.require('Blockly.utils.array');
 const Extensions = goog.require('Blockly.Extensions');
 const Tooltip = goog.require('Blockly.Tooltip');
+const arrayUtils = goog.require('Blockly.utils.array');
 const common = goog.require('Blockly.common');
 const constants = goog.require('Blockly.constants');
 const eventUtils = goog.require('Blockly.Events.utils');
@@ -27,7 +27,6 @@ const fieldRegistry = goog.require('Blockly.fieldRegistry');
 const idGenerator = goog.require('Blockly.utils.idGenerator');
 const object = goog.require('Blockly.utils.object');
 const parsing = goog.require('Blockly.utils.parsing');
-const utils = goog.require('Blockly.utils');
 const {ASTNode} = goog.require('Blockly.ASTNode');
 const {Blocks} = goog.require('Blockly.blocks');
 /* eslint-disable-next-line no-unused-vars */
@@ -870,7 +869,7 @@ Block.prototype.isDuplicatable = function() {
     return true;
   }
   return this.workspace.isCapacityAvailable(
-      utils.getBlockTypeCounts(this, true));
+      common.getBlockTypeCounts(this, true));
 };
 
 /**
